+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+### Added
+- Vertical orientation for volume sliders is now available via `-b v` or `--bar vertical`.
+
 ### Fixed
 - (CSS) Icon wasn't affected by style changes in `default.css`, because class selector was invalid.